# VideoGameApiWithKafka/Video Game API 1.9/openapi.yaml
openapi: 3.0.0
info:
  title: Video Game Exchange
  version: 1.0.0
servers:
- url: http://localhost:8080
paths:
  /users:
    get:
      operationId: listUsers
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
          description: A paged array of Users
      summary: List All Users
      tags:
      - users
      x-accepts: application/json
      x-tags:
      - tag: users
    post:
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "201":
          description: Null response
        default:
          description: unexpected error
      summary: Create a User
      tags:
      - users
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: users
  /users/{userId}:
    delete:
      operationId: deleteUser
      parameters:
      - description: The id of the user to delete
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Null response
        default:
          description: unexpected error
      summary: Delete one User
      tags:
      - users
      x-accepts: application/json
      x-tags:
      - tag: users
    get:
      operationId: getOneUser
      parameters:
      - description: The id of the user to get
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: One User Object
      summary: Get One User
      tags:
      - users
      x-accepts: application/json
      x-tags:
      - tag: users
    patch:
      operationId: updateSomeUser
      parameters:
      - description: The id of the user to update
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "204":
          description: Null response
        default:
          description: unexpected error
      summary: Update Some User Attributes
      tags:
      - users
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: users
    put:
      operationId: updateUser
      parameters:
      - description: The id of the user to update
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "204":
          description: Null response
        default:
          description: unexpected error
      summary: Update all User Attributes
      tags:
      - users
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: users
  /users/{userId}/{gameId}:
    post: 
      parameters:
      - description: The id of the user to add game to
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      - description: The id of the game to add to user
        explode: false
        in: path
        name: gameId
        required: true
        schema:
          type: string
        style: simple
      operationId: addGameUser
      responses:
        "201":
          description: Null response
        default:
          description: unexpected error
      summary: Add a Game to a User
      tags:
      - users
      - games
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: users
    delete: 
      parameters:
      - description: The id of the user to delete game from
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      - description: The id of the game to delete from user
        explode: false
        in: path
        name: gameId
        required: true
        schema:
          type: string
        style: simple
      operationId: deleteGameUser
      responses:
        "201":
          description: Null response
        default:
          description: unexpected error
      summary: Remove a Game from a User  
      tags:
      - users
      - games
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: users
...
components:
  schemas:
    User:
      example:
        zip: zip
        address: address
        city: city
        name: name
        games:
        - null
        - null
        id: 0
        state: state
        email: email
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        games:
          items:
            type: string
          type: array
      required:
      - address
      - city
      - email
      - id
      - name
      - state
      - zip
      type: object
...